+++ R/metadata-cache.R
@@ -245,7 +245,7 @@ cranlike_metadata_cache <- R6Class(
     ## We use this to make sure that different versions of pkgcache can
     ## share the same metadata cache directory. It is used to calculate
     ## the hash of the cached RDS file.
-    cache_version = "5",
+    cache_version = "6",
 
     data = NULL,
     data_time = NULL,