+++ tests/testthat/test-packages-gz.R
@@ -49,8 +49,8 @@ test_that("packages_make_sources", {
     packages_make_sources(
       "URL", "source", c("s/c/xx.tar.gz", "s/c/yy.tar.gz"), "s/c",
       c("p1", "p2"), c("1.0", "2.0"), type = "cran", downloadurl = NULL),
-    list(c("URL/s/c/xx.tar.gz", "URL/s/c/Archive/p1_1.0.tar.gz"),
-         c("URL/s/c/yy.tar.gz", "URL/s/c/Archive/p2_2.0.tar.gz"))
+    list(c("URL/s/c/xx.tar.gz", "URL/s/c/Archive/p1/p1_1.0.tar.gz"),
+         c("URL/s/c/yy.tar.gz", "URL/s/c/Archive/p2/p2_2.0.tar.gz"))
   )
 
   urls <- c("foo", NA)